backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
scipy
sqlalchemy
python-jose
pytest